Fix sign-compare in asserts and missing .data() calls
The testrunner builds with different flags (no NDEBUG), exposing
sign-compare warnings in assert() expressions that use .size().
Add static_cast<int>() around .size() in asserts in status.cpp,
analyze.cpp, and pragma.cpp.

Fix missing .data() on aLck in os_unix.cpp line 4321.

+++ src/analyze.cpp
@@ -244,7 +244,7 @@ static void openStatTable(
 
   /* Open the sqlite_stat[134] tables for writing. */
   for(i=0; i<nToOpen; i++){
-    assert( i<aTable.size() );
+    assert( i<static_cast<int>(aTable.size()) );
     sqlite3VdbeAddOp4Int(v, OP_OpenWrite, iStatCur+i, (int)aRoot[i], iDb, 3);
     sqlite3VdbeChangeP5(v, aCreateTbl[i]);
     VdbeComment((v, aTable[i].zName));
